Javafx module export error fixed
Earlier plan was to migrate the project into JPMS (Java Platform Module
System) which is a java 11 feature. But many difficulty was faced while
trying to migrate.

So we are temporarily overriding the run arguments to export the javafx
modules (such as javafx.controls) to unnamed modules. This tweak doesnot
throw a runtime exception.

Signed-off-by: Rajdeep Roy Chowdhury <rrajdeeproychowdhury@gmail.com>